helperscripts/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.1.0
cachetools==5.3.0
certifi==2022.12.7
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==7.2.3
cssselect==1.2.0
decorator==5.1.1
distlib==0.3.6
dnspython==2.3.0
docker==6.0.1
duckduckgo-search==2.8.6
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.11.0
flake8==6.0.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
gTTS==2.3.1
h11==0.14.0
httplib2==0.22.0
identify==2.5.22
idna==3.4
iniconfig==2.0.0
ipython==8.11.0
isort==5.12.0
jedi==0.18.2
jsonschema==4.17.3
loguru==0.7.0
lxml==4.9.2
matplotlib-inline==0.1.6
matplotlib==3.7.1
mccabe==0.7.0
MouseInfo==0.1.3
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.7.0
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.2
orjson==3.8.10
outcome==1.2.0
packaging==23.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.1.1
playsound==1.2.2
pluggy==1.0.0
pre-commit==3.2.2
prompt-toolkit==3.0.38
protobuf==4.22.3
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycodestyle==2.10.0
pyflakes==3.0.1
PyGetWindow==0.0.9
Pygments==2.14.0
PyMsgBox==1.0.9
pynput==1.7.6
pyobjc==9.0.1
pyobjc-core==9.0.1
pyobjc-framework-Accessibility==9.0.1
pyobjc-framework-Accounts==9.0.1
pyobjc-framework-AddressBook==9.0.1
pyobjc-framework-AdServices==9.0.1
pyobjc-framework-AdSupport==9.0.1
pyobjc-framework-AppleScriptKit==9.0.1
pyobjc-framework-AppleScriptObjC==9.0.1
pyobjc-framework-ApplicationServices==9.0.1
pyobjc-framework-AppTrackingTransparency==9.0.1
pyobjc-framework-AudioVideoBridging==9.0.1
pyobjc-framework-AuthenticationServices==9.0.1
pyobjc-framework-AutomaticAssessmentConfiguration==9.0.1
pyobjc-framework-Automator==9.0.1
pyobjc-framework-AVFoundation==9.0.1
pyobjc-framework-AVKit==9.0.1
pyobjc-framework-BusinessChat==9.0.1
pyobjc-framework-CalendarStore==9.0.1
pyobjc-framework-CallKit==9.0.1
pyobjc-framework-CFNetwork==9.0.1
pyobjc-framework-ClassKit==9.0.1
pyobjc-framework-CloudKit==9.0.1
pyobjc-framework-Cocoa==9.0.1
pyobjc-framework-Collaboration==9.0.1
pyobjc-framework-ColorSync==9.0.1
pyobjc-framework-Contacts==9.0.1
pyobjc-framework-ContactsUI==9.0.1
pyobjc-framework-CoreAudio==9.0.1
pyobjc-framework-CoreAudioKit==9.0.1
pyobjc-framework-CoreBluetooth==9.0.1
pyobjc-framework-CoreData==9.0.1
pyobjc-framework-CoreHaptics==9.0.1
pyobjc-framework-CoreLocation==9.0.1
pyobjc-framework-CoreMedia==9.0.1
pyobjc-framework-CoreMediaIO==9.0.1
pyobjc-framework-CoreMIDI==9.0.1
pyobjc-framework-CoreML==9.0.1
pyobjc-framework-CoreMotion==9.0.1
pyobjc-framework-CoreServices==9.0.1
pyobjc-framework-CoreSpotlight==9.0.1
pyobjc-framework-CoreText==9.0.1
pyobjc-framework-CoreWLAN==9.0.1
pyobjc-framework-CryptoTokenKit==9.0.1
pyobjc-framework-DataDetection==9.0.1
pyobjc-framework-DeviceCheck==9.0.1
pyobjc-framework-DictionaryServices==9.0.1
pyobjc-framework-DiscRecording==9.0.1
pyobjc-framework-DiscRecordingUI==9.0.1
pyobjc-framework-DiskArbitration==9.0.1
pyobjc-framework-DVDPlayback==9.0.1
pyobjc-framework-EventKit==9.0.1
pyobjc-framework-ExceptionHandling==9.0.1
pyobjc-framework-ExecutionPolicy==9.0.1
pyobjc-framework-ExternalAccessory==9.0.1
pyobjc-framework-FileProvider==9.0.1
pyobjc-framework-FileProviderUI==9.0.1
pyobjc-framework-FinderSync==9.0.1
pyobjc-framework-FSEvents==9.0.1
pyobjc-framework-GameCenter==9.0.1
pyobjc-framework-GameController==9.0.1
pyobjc-framework-GameKit==9.0.1
pyobjc-framework-GameplayKit==9.0.1
pyobjc-framework-ImageCaptureCore==9.0.1
pyobjc-framework-IMServicePlugIn==9.0.1
pyobjc-framework-InputMethodKit==9.0.1
pyobjc-framework-InstallerPlugins==9.0.1
pyobjc-framework-InstantMessage==9.0.1
pyobjc-framework-Intents==9.0.1
pyobjc-framework-IntentsUI==9.0.1
pyobjc-framework-IOSurface==9.0.1
pyobjc-framework-iTunesLibrary==9.0.1
pyobjc-framework-KernelManagement==9.0.1
pyobjc-framework-LatentSemanticMapping==9.0.1
pyobjc-framework-LaunchServices==9.0.1
pyobjc-framework-libdispatch==9.0.1
pyobjc-framework-LinkPresentation==9.0.1
pyobjc-framework-LocalAuthentication==9.0.1
pyobjc-framework-LocalAuthenticationEmbeddedUI==9.0.1
pyobjc-framework-MailKit==9.0.1
pyobjc-framework-MapKit==9.0.1
pyobjc-framework-MediaAccessibility==9.0.1
pyobjc-framework-MediaLibrary==9.0.1
pyobjc-framework-MediaPlayer==9.0.1
pyobjc-framework-MediaToolbox==9.0.1
pyobjc-framework-Metal==9.0.1
pyobjc-framework-MetalKit==9.0.1
pyobjc-framework-MetalPerformanceShaders==9.0.1
pyobjc-framework-MetalPerformanceShadersGraph==9.0.1
pyobjc-framework-MetricKit==9.0.1
pyobjc-framework-MLCompute==9.0.1
pyobjc-framework-ModelIO==9.0.1
pyobjc-framework-MultipeerConnectivity==9.0.1
pyobjc-framework-NaturalLanguage==9.0.1
pyobjc-framework-NetFS==9.0.1
pyobjc-framework-Network==9.0.1
pyobjc-framework-NetworkExtension==9.0.1
pyobjc-framework-NotificationCenter==9.0.1
pyobjc-framework-OpenDirectory==9.0.1
pyobjc-framework-OSAKit==9.0.1
pyobjc-framework-OSLog==9.0.1
pyobjc-framework-PassKit==9.0.1
pyobjc-framework-PencilKit==9.0.1
pyobjc-framework-Photos==9.0.1
pyobjc-framework-PhotosUI==9.0.1
pyobjc-framework-PreferencePanes==9.0.1
pyobjc-framework-PushKit==9.0.1
pyobjc-framework-Quartz==9.0.1
pyobjc-framework-QuickLookThumbnailing==9.0.1
pyobjc-framework-ReplayKit==9.0.1
pyobjc-framework-SafariServices==9.0.1
pyobjc-framework-SceneKit==9.0.1
pyobjc-framework-ScreenCaptureKit==9.0.1
pyobjc-framework-ScreenSaver==9.0.1
pyobjc-framework-ScreenTime==9.0.1
pyobjc-framework-ScriptingBridge==9.0.1
pyobjc-framework-SearchKit==9.0.1
pyobjc-framework-Security==9.0.1
pyobjc-framework-SecurityFoundation==9.0.1
pyobjc-framework-SecurityInterface==9.0.1
pyobjc-framework-ServiceManagement==9.0.1
pyobjc-framework-ShazamKit==9.0.1
pyobjc-framework-Social==9.0.1
pyobjc-framework-SoundAnalysis==9.0.1
pyobjc-framework-Speech==9.0.1
pyobjc-framework-SpriteKit==9.0.1
pyobjc-framework-StoreKit==9.0.1
pyobjc-framework-SyncServices==9.0.1
pyobjc-framework-SystemConfiguration==9.0.1
pyobjc-framework-SystemExtensions==9.0.1
pyobjc-framework-UniformTypeIdentifiers==9.0.1
pyobjc-framework-UserNotifications==9.0.1
pyobjc-framework-UserNotificationsUI==9.0.1
pyobjc-framework-VideoSubscriberAccount==9.0.1
pyobjc-framework-VideoToolbox==9.0.1
pyobjc-framework-Virtualization==9.0.1
pyobjc-framework-Vision==9.0.1
pyobjc-framework-WebKit==9.0.1
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
PySocks==1.7.1
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-benchmark==4.0.0
pytest-cov==4.0.0
pytest-integration==0.2.3
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.4
pywatchman==1.4.1
PyYAML==6.0
readability-lxml==0.8.1
redis==4.5.4
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
rubicon-objc==0.4.5
selenium==4.8.3
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sourcery==1.2.0
stack-data==0.6.2
tiktoken==0.3.3
tomli==2.0.1
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.13.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0
wcwidth==0.2.6
webdriver-manager==3.8.6
websocket-client==1.5.1
wsproto==1.2.0
yarl==1.8.2
